moved version logging to `Libraries`
